Added terminate_job_when_celery_is_down functionality
Former-commit-id: 4f1752799a6e5895d805c3f673636d37e3c6a951 [formerly 4f1752799a6e5895d805c3f673636d37e3c6a951 [formerly 7d234b11cc76944ce219c7df36f53a4813141855]]
Former-commit-id: 28110f3e251e2c5cd291322a744ab00b8272093b
Former-commit-id: b90ce15eb72bbfc5693ef9dce6c377183d13df55

+++ openquake/engine/celery_node_monitor.py
@@ -24,6 +24,9 @@ import threading
 
 import celery.task.control
 
+from openquake.engine import logs
+from openquake.engine.utils import config, general
+
 
 class MasterKilled(KeyboardInterrupt):
     """
@@ -107,9 +110,13 @@ class CeleryNodeMonitor(object):
         while self.job_is_running(sleep=self.interval):
             live_nodes = set(celery.task.control.inspect().ping() or {})
             if live_nodes < self.live_nodes:
-                print >> sys.stderr, 'Cluster nodes not accessible: %s' % (
-                    self.live_nodes - live_nodes)
-                os.kill(os.getpid(), signal.SIGABRT)  # commit suicide
+                dead_nodes = list(self.live_nodes - live_nodes)
+                logs.LOG.critical(
+                    'Cluster nodes not accessible: %s', dead_nodes)
+                terminate = general.str2bool(
+                    config.get('celery', 'terminate_job_when_celery_is_down'))
+                if terminate:
+                    os.kill(os.getpid(), signal.SIGABRT)  # commit suicide
 
     def job_is_running(self, sleep):
         """
